refactor: move render paths renderer
