+++ svg-time-series/src/chart/render.ts
@@ -1,6 +1,5 @@
 import { ScaleLinear, ScaleTime } from "d3-scale";
 import { BaseType, Selection } from "d3-selection";
-import { line } from "d3-shape";
 
 import { MyAxis, Orientation } from "../axis.ts";
 import { ViewportTransform } from "../ViewportTransform.ts";
@@ -122,24 +121,6 @@ export function setupRender(
   const dimensions: Dimensions = { width, height };
 
   return { scales, axes, paths, transforms, dimensions };
-}
-
-export function renderPaths(
-  state: RenderState,
-  dataArr: Array<[number, number?]>,
-) {
-  const drawLine = (cityIdx: number) =>
-    line<[number, number?]>()
-      .defined((d) => {
-        return !(isNaN(d[cityIdx]!) || d[cityIdx] == null);
-      })
-      .x((d, i) => i)
-      .y((d) => d[cityIdx]!);
-
-  state.paths.path.attr(
-    "d",
-    (cityIndex: number) => drawLine(cityIndex)(dataArr) ?? "",
-  );
 }
 
 export function refreshChart(state: RenderState, data: ChartData) {
