+++ svg-time-series/src/chart/render/paths.ts
@@ -1,5 +1,7 @@
 import { BaseType, Selection } from "d3-selection";
+import { line } from "d3-shape";
 import { ViewportTransform } from "../../ViewportTransform.ts";
+import type { RenderState } from "../render.ts";
 
 export interface PathSet {
   path: Selection<SVGPathElement, number, any, unknown>;
@@ -36,4 +38,20 @@ export function createTransforms(paths: PathSet): TransformPair {
     sf = new ViewportTransform();
   }
   return { ny, sf };
+}
+
+export function renderPaths(
+  state: RenderState,
+  dataArr: Array<[number, number?]>,
+) {
+  const drawLine = (cityIdx: number) =>
+    line<[number, number?]>()
+      .defined((d) => !(isNaN(d[cityIdx]!) || d[cityIdx] == null))
+      .x((_, i) => i)
+      .y((d) => d[cityIdx]!);
+
+  state.paths.path.attr(
+    "d",
+    (cityIndex: number) => drawLine(cityIndex)(dataArr) ?? "",
+  );
 }
